+++ demo/pages/hook/collections.tsx
@@ -3,6 +3,7 @@ import { useCollections } from '@reservoir0x/reservoir-kit-ui'
 import { ConnectButton } from '@rainbow-me/rainbowkit'
 import { useInView } from 'react-intersection-observer'
 import { useEffect } from 'react'
+import ChainSwitcher from 'components/ChainSwitcher'
 
 const Collections: NextPage = () => {
   const { data: collections, fetchNextPage, hasNextPage } = useCollections({})
@@ -48,6 +49,7 @@ const Collections: NextPage = () => {
       ) : (
         <div>No more data</div>
       )}
+      <ChainSwitcher style={{ right: 16 }} />
     </div>
   )
 }