+++ demo/pages/modal/buy.tsx
@@ -5,6 +5,7 @@ import ThemeSwitcher from 'components/ThemeSwitcher'
 import { useState } from 'react'
 import DeeplinkCheckbox from 'components/DeeplinkCheckbox'
 import { useRouter } from 'next/router'
+import ChainSwitcher from 'components/ChainSwitcher'
 
 const DEFAULT_COLLECTION_ID =
   process.env.NEXT_PUBLIC_DEFAULT_COLLECTION_ID ||
@@ -166,6 +167,7 @@ const BuyPage: NextPage = () => {
         }}
       />
       <ThemeSwitcher />
+      <ChainSwitcher />
     </div>
   )
 }