+++ packages/ui/src/modal/buy/BuyModal.tsx
@@ -72,6 +72,7 @@ type Props = Pick<Parameters<typeof Modal>['0'], 'trigger'> & {
   normalizeRoyalties?: boolean
   copyOverrides?: Partial<typeof ModalCopy>
   walletClient?: ReservoirWallet | WalletClient
+  usePermit?: boolean
   onConnectWallet: () => void
   onGoToToken?: () => any
   onPurchaseComplete?: (data: PurchaseData) => void
@@ -114,6 +115,7 @@ export function BuyModal({
   defaultQuantity,
   copyOverrides,
   walletClient,
+  usePermit,
   onConnectWallet,
   onPurchaseComplete,
   onPurchaseError,
@@ -148,6 +150,7 @@ export function BuyModal({
       feesOnTopUsd={feesOnTopUsd}
       normalizeRoyalties={normalizeRoyalties}
       walletClient={walletClient}
+      usePermit={usePermit}
       onConnectWallet={onConnectWallet}
     >
       {({
@@ -224,7 +227,7 @@ export function BuyModal({
 
         const failedPurchases = quantity - totalPurchases
         const successfulPurchases = quantity - failedPurchases
-        const finalTxHash = lastStepItems[lastStepItems.length - 1]?.txHash
+        const finalTxHashes = lastStepItems[lastStepItems.length - 1]?.txHashes
 
         const price =
           totalPrice || BigInt(token?.token?.lastSale?.price?.amount?.raw || 0)
@@ -504,13 +507,13 @@ export function BuyModal({
                 {stepData && (
                   <Progress
                     title={stepData?.currentStep.action || ''}
-                    txHash={stepData?.currentStepItem.txHash}
-                    blockExplorerBaseUrl={`${blockExplorerBaseUrl}/tx/${stepData?.currentStepItem.txHash}`}
+                    txHashes={stepData?.currentStepItem.txHashes}
+                    blockExplorerBaseUrl={blockExplorerBaseUrl}
                   />
                 )}
                 <Button disabled={true} css={{ m: '$4' }}>
                   <Loader />
-                  {stepData?.currentStepItem.txHash
+                  {stepData?.currentStepItem?.txHashes
                     ? copy.ctaAwaitingValidation
                     : copy.ctaAwaitingApproval}
                 </Button>
@@ -577,21 +580,29 @@ export function BuyModal({
                   )}
                   {totalPurchases > 1 && (
                     <Flex direction="column" css={{ gap: '$2' }}>
-                      {stepData?.currentStep.items?.map((item) => {
-                        const txHash = item.txHash
-                          ? `${truncateAddress(item.txHash)}`
-                          : ''
-                        return (
-                          <Anchor
-                            href={`${blockExplorerBaseUrl}/tx/${item?.txHash}`}
-                            color="primary"
-                            weight="medium"
-                            target="_blank"
-                            css={{ fontSize: 12 }}
-                          >
-                            View transaction: {txHash}
-                          </Anchor>
-                        )
+                      {stepData?.currentStep?.items?.map((item, itemIndex) => {
+                        if (
+                          Array.isArray(item?.txHashes) &&
+                          item?.txHashes.length > 0
+                        ) {
+                          return item.txHashes.map((txHash, txHashIndex) => {
+                            const truncatedTxHash = truncateAddress(txHash)
+                            return (
+                              <Anchor
+                                key={`${itemIndex}-${txHashIndex}`}
+                                href={`${blockExplorerBaseUrl}/tx/${txHash}`}
+                                color="primary"
+                                weight="medium"
+                                target="_blank"
+                                css={{ fontSize: 12 }}
+                              >
+                                View transaction: {truncatedTxHash}
+                              </Anchor>
+                            )
+                          })
+                        } else {
+                          return null
+                        }
                       })}
                     </Flex>
                   )}
@@ -633,15 +644,28 @@ export function BuyModal({
                           Your transaction went through successfully
                         </Text>
                       </Flex>
-                      <Anchor
-                        color="primary"
-                        weight="medium"
-                        css={{ fontSize: 12 }}
-                        href={`${blockExplorerBaseUrl}/tx/${finalTxHash}`}
-                        target="_blank"
+
+                      <Flex
+                        direction="column"
+                        align="center"
+                        css={{ gap: '$2' }}
                       >
-                        View on {blockExplorerBaseName}
-                      </Anchor>
+                        {finalTxHashes?.map((txHash, index) => {
+                          const truncatedTxHash = truncateAddress(txHash)
+                          return (
+                            <Anchor
+                              key={index}
+                              href={`${blockExplorerBaseUrl}/tx/${txHash}`}
+                              color="primary"
+                              weight="medium"
+                              target="_blank"
+                              css={{ fontSize: 12 }}
+                            >
+                              View transaction: {truncatedTxHash}
+                            </Anchor>
+                          )
+                        })}
+                      </Flex>
                     </>
                   )}
                 </Flex>