+++ packages/ui/src/modal/buy/BuyModalRenderer.tsx
@@ -5,6 +5,7 @@ import React, {
   useCallback,
   ReactNode,
   useMemo,
+  useContext,
 } from 'react'
 import {
   useTokens,
@@ -28,6 +29,7 @@ import * as allChains from 'viem/chains'
 import usePaymentTokens, {
   EnhancedCurrency,
 } from '../../hooks/usePaymentTokens'
+import { ProviderOptionsContext } from '../../ReservoirKitProvider'
 
 type Item = Parameters<ReservoirClientActions['buyToken']>['0']['items'][0]
 
@@ -100,6 +102,7 @@ type Props = {
   onConnectWallet: () => void
   children: (props: ChildrenProps) => ReactNode
   walletClient?: ReservoirWallet | WalletClient
+  usePermit?: boolean
 }
 
 export const BuyModalRenderer: FC<Props> = ({
@@ -115,6 +118,7 @@ export const BuyModalRenderer: FC<Props> = ({
   onConnectWallet,
   children,
   walletClient,
+  usePermit,
 }) => {
   const [totalPrice, setTotalPrice] = useState(0n)
   const [totalIncludingFees, setTotalIncludingFees] = useState(0n)
@@ -131,6 +135,7 @@ export const BuyModalRenderer: FC<Props> = ({
 
   const client = useReservoirClient()
   const currentChain = client?.currentChain()
+  const providerOptionsContext = useContext(ProviderOptionsContext)
 
   const rendererChain = chainId
     ? client?.chains.find(({ id }) => id === chainId) || currentChain
@@ -170,7 +175,9 @@ export const BuyModalRenderer: FC<Props> = ({
   )
 
   const paymentCurrency = paymentTokens?.find(
-    (paymentToken) => paymentToken?.address === _paymentCurrency?.address
+    (paymentToken) =>
+      paymentToken?.address === _paymentCurrency?.address &&
+      paymentToken?.chainId === _paymentCurrency?.chainId
   )
 
   const { data: tokens, mutate: mutateTokens } = useTokens(
@@ -271,6 +278,7 @@ export const BuyModalRenderer: FC<Props> = ({
       onlyPath: true,
       partial: true,
       currency: paymentCurrency?.address,
+      currencyChainId: paymentCurrency?.chainId,
     }
 
     if (normalizeRoyalties !== undefined) {
@@ -321,40 +329,87 @@ export const BuyModalRenderer: FC<Props> = ({
     normalizeRoyalties,
     rendererChain,
     paymentCurrency?.address,
+    paymentCurrency?.chainId,
   ])
 
   useEffect(() => {
-    fetchPath()
-  }, [fetchPath])
+    // ensure the tokens api has been fetched first
+    if (token) {
+      fetchPath()
+    }
+  }, [fetchPath, token])
+
+  const getCurrencyDetails = useCallback(() => {
+    if (orderId) {
+      return {
+        currency: listing?.price?.currency?.contract,
+        decimals: listing?.price?.currency?.decimals,
+        symbol: listing?.price?.currency?.symbol,
+        chainId: rendererChain?.id || 1,
+      }
+    } else if (is1155) {
+      return {
+        currency: path?.[0]?.currency,
+        decimals: path?.[0]?.currencyDecimals,
+        symbol: path?.[0]?.currencySymbol,
+        chainId: rendererChain?.id || 1,
+      }
+    } else {
+      return {
+        currency: token?.market?.floorAsk?.price?.currency?.contract,
+        decimals: token?.market?.floorAsk?.price?.currency?.decimals,
+        symbol: token?.market?.floorAsk?.price?.currency?.symbol,
+        chainId: rendererChain?.id || 1,
+      }
+    }
+  }, [orderId, is1155, listing, path, token, rendererChain])
 
   useEffect(() => {
-    if (!paymentTokens[0] || paymentCurrency) {
+    if (
+      !paymentTokens[0] ||
+      paymentCurrency ||
+      !token ||
+      (orderId && !listing) ||
+      (is1155 && !path)
+    ) {
       return
     }
 
-    let selectedCurrency
+    const { currency, decimals, symbol, chainId } = getCurrencyDetails()
 
-    if (orderId) {
-      selectedCurrency =
-        paymentTokens.find(
-          (token) =>
-            token.address === listing?.price?.currency?.contract?.toLowerCase()
-        ) || paymentTokens[0]
-    } else if (is1155) {
-      selectedCurrency =
-        paymentTokens.find(
-          (token) => token.address === path?.[0].currency?.toLowerCase()
-        ) || paymentTokens[0]
+    // Determine whether to include the listing currency unconditionally
+    const includeListingCurrency =
+      providerOptionsContext.alwaysIncludeListingCurrency !== false
+
+    let selectedCurrency
+    if (includeListingCurrency && currency) {
+      selectedCurrency = {
+        address: currency.toLowerCase() as Address,
+        decimals: decimals || 18,
+        symbol: symbol || '',
+        chainId: chainId,
+      }
     } else {
-      selectedCurrency =
-        paymentTokens.find(
-          (paymentToken) =>
-            paymentToken.address ===
-            token?.market?.floorAsk?.price?.currency?.contract?.toLowerCase()
-        ) || paymentTokens[0]
+      selectedCurrency = paymentTokens.find(
+        (token) => token.address === currency?.toLowerCase()
+      )
+
+      if (!selectedCurrency) {
+        selectedCurrency = paymentTokens[0]
+      }
     }
+
     setPaymentCurrency(selectedCurrency)
-  }, [paymentTokens, chainCurrency])
+  }, [
+    paymentTokens,
+    paymentCurrency,
+    orderId,
+    is1155,
+    listing,
+    path,
+    token,
+    providerOptionsContext.alwaysIncludeListingCurrency,
+  ])
 
   const buyToken = useCallback(async () => {
     if (!wallet) {
@@ -367,13 +422,16 @@ export const BuyModalRenderer: FC<Props> = ({
     }
 
     let activeWalletChain = getNetwork().chain
-    if (activeWalletChain && rendererChain?.id !== activeWalletChain?.id) {
+    if (
+      activeWalletChain &&
+      paymentCurrency?.chainId !== activeWalletChain?.id
+    ) {
       activeWalletChain = await switchNetwork({
-        chainId: rendererChain?.id as number,
+        chainId: paymentCurrency?.chainId as number,
       })
     }
 
-    if (rendererChain?.id !== activeWalletChain?.id) {
+    if (paymentCurrency?.chainId !== activeWalletChain?.id) {
       const error = new Error(`Mismatching chainIds`)
       setTransactionError(error)
       throw error
@@ -395,6 +453,7 @@ export const BuyModalRenderer: FC<Props> = ({
 
     let options: BuyTokenOptions = {
       currency: paymentCurrency?.address,
+      currencyChainId: paymentCurrency?.chainId,
     }
 
     if (feesOnTopBps && feesOnTopBps?.length > 0) {
@@ -423,6 +482,10 @@ export const BuyModalRenderer: FC<Props> = ({
 
     if (normalizeRoyalties !== undefined) {
       options.normalizeRoyalties = normalizeRoyalties
+    }
+
+    if (usePermit) {
+      options.usePermit = true
     }
 
     setBuyStep(BuyStep.Approving)
@@ -535,6 +598,8 @@ export const BuyModalRenderer: FC<Props> = ({
     totalIncludingFees,
     wallet,
     paymentCurrency?.address,
+    paymentCurrency?.chainId,
+    usePermit,
     mutateListings,
     mutateTokens,
     mutateCollection,
@@ -694,6 +759,7 @@ export const BuyModalRenderer: FC<Props> = ({
       setQuantity(1)
       setPath(undefined)
       setPaymentCurrency(undefined)
+      setToken(undefined)
     } else {
       setQuantity(defaultQuantity || 1)
     }